requirements.txt
absl-py==0.2.0
aiofiles==0.3.2
aiohttp==3.3.0
appnope==0.1.0
astor==0.6.2
async-timeout==2.0.1
atomicwrites==1.1.5
attrs==18.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.6.17
botocore==1.9.17
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
coverage==4.4.2
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.2.1
dill==0.2.7.1
docutils==0.14
entrypoints==0.2.3
ftfy==4.4.3
gast==0.2.0
gensim==3.4.0
grpcio==1.12.0
h5py==2.8.0
html5lib==0.9999999
httptools==0.0.10
idna==2.6
idna-ssl==1.0.1
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.31.12
jupyterlab-launcher==0.10.5
Keras==2.1.6
kiwisolver==1.0.1
lxml==4.2.3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
more-itertools==4.2.0
msgpack-numpy==0.4.1
msgpack-python==0.5.4
multidict==4.3.1
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
nose-timer==0.7.1
notebook==5.4.1
notedown==1.5.1
numpy==1.14.3
pandas==0.23.3
pandoc-attributes==0.1.7
pandocfilters==1.4.2
parso==0.1.1
pathlib==1.0.1
pexpect==4.4.0
pickleshare==0.7.4
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
progressbar2==3.36.0
progressbar33==2.4
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
py==1.5.4
pyasn1==0.4.2
Pygments==2.2.0
PyMemoize==1.0.2
pyparsing==2.2.0
pytest==3.6.2
pytest-cov==2.5.1
pytest-sugar==0.9.1
python-dateutil==2.6.1
python-utils==2.3.0
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
psycopg2==2.7.5
rednose==1.2.3
regex==2017.4.5
requests==2.18.4
s3transfer==0.1.13
sanic==0.7.0
sanic-compress==0.1.1
Sanic-Cors==0.9.4
Sanic-Plugins-Framework==0.5.2.dev20180201
scipy==1.0.0
schedule==0.5.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.7
spacy==2.0.10
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.1
termstyle==0.1.11
testpath==0.3.1
thinc==6.10.2
toolz==0.9.0
tornado==5.0.1
tqdm==4.20.0
traitlets==4.3.2
ujson==1.35
urllib3==1.22
uvloop==0.9.1
wcwidth==0.1.7
webencodings==0.5.1
websockets==4.0.1
Werkzeug==0.14.1
wrapt==1.10.11
yarl==1.2.4
plotly==3.0.0
isoweek==1.3.3
Tinys3==0.1.12
python-slugify==1.2.5
timeout-decorator==0.4.0
